requirements.txt
Django==4.2
gunicorn==21.2.0
dj_database_url==2.0.0
django_crispy_forms==2.10
django-crispy-bootstrap4